src/mcp_servers/gemini/requirements.txt
google-generativeai
dotenv
mcp
Pillow
pydantic